se cambio a Long el id de avisos, y se corrigió bug, que no permitia borrar los avisos al cambiar a estado distinto de terminado

+++ sistema-integral/src/main/java/com/megasolution/app/sistemaintegral/avisos/models/repositories/IAvisoRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IAvisoRepository extends JpaRepository<Aviso, Integer>{
+public interface IAvisoRepository extends JpaRepository<Aviso, Long>{
     
     @Query(value = "SELECT COUNT(*) FROM avisos;", nativeQuery = true)
     public Integer contarAvisos();
